Incorporate the new `locks` module, and deprecate lock removal functions
