+++ conans/test/functional/old/package_integrity_test.py
@@ -9,25 +9,6 @@ from conans.util.files import set_dirty
 
 class PackageIngrityTest(unittest.TestCase):
 
-    def remove_locks_test(self):
-        client = TestClient()
-        client.save({"conanfile.py": GenConanfile().with_name("Hello").with_version("0.1")})
-        client.run("create . lasote/testing")
-        self.assertNotIn('does not contain a number!', client.out)
-        ref = ConanFileReference.loads("Hello/0.1@lasote/testing")
-        conan_folder = client.cache.package_layout(ref).base_folder()
-        self.assertIn("locks", os.listdir(conan_folder))
-        self.assertTrue(os.path.exists(conan_folder + ".count"))
-        self.assertTrue(os.path.exists(conan_folder + ".count.lock"))
-        client.run("remove * --locks", assert_error=True)
-        self.assertIn("ERROR: Specifying a pattern is not supported", client.out)
-        client.run("remove", assert_error=True)
-        self.assertIn('ERROR: Please specify a pattern to be removed ("*" for all)', client.out)
-        client.run("remove --locks")
-        self.assertNotIn("locks", os.listdir(conan_folder))
-        self.assertFalse(os.path.exists(conan_folder + ".count"))
-        self.assertFalse(os.path.exists(conan_folder + ".count.lock"))
-
     def upload_dirty_test(self):
         test_server = TestServer([], users={"lasote": "mypass"})
         client = TestClient(servers={"default": test_server},
